added blu sim gui
